+++ node/node.go
@@ -386,6 +386,9 @@ type Node struct {
 
 	resourceManager resource.Manager
 
+	// Tracks bootstrapping of subnets
+	subnets *chains.Subnets
+
 	// Tracks the CPU/disk usage caused by processing
 	// messages of each peer.
 	resourceTracker tracker.ResourceTracker
@@ -1058,6 +1061,28 @@ func (n *Node) addDefaultVMAliases() error {
 	return nil
 }
 
+func (n *Node) reportUnhealthy(unhealthy bool) {
+	subnetsLeftToBootstrap := n.subnets.Bootstrapping()
+	if len(subnetsLeftToBootstrap) > 0 {
+		n.Log.Verbo("Still bootstrapping, will not change logging level",
+			zap.Int("subnets left to bootstrap", len(subnetsLeftToBootstrap)))
+		return
+	}
+
+	// Either amplify the log level or revert it to what it was.
+	// It is assumed that reportUnhealthy() is called upon every health check
+	// iteration, otherwise there is nothing to reset the logging level to what it was
+	// when the health check succeeds after failing.
+	// Similarly, if the health check continuously fails, this function needs to be called
+	// repeatedly, otherwise there is nothing to revert the logging level to what it was
+	// once it was amplified for too long.
+	if unhealthy {
+		n.LogFactory.Amplify()
+	} else {
+		n.LogFactory.Revert()
+	}
+}
+
 // Create the chainManager and register the following VMs:
 // AVM, Simple Payments DAG, Simple Payments Chain, and Platform VM
 // Assumes n.DBManager, n.vdrs all initialized (non-nil)
@@ -1129,6 +1154,7 @@ func (n *Node) initChainManager(avaxAssetID ids.ID) error {
 	if err != nil {
 		return fmt.Errorf("failed to initialize subnets: %w", err)
 	}
+	n.subnets = subnets
 
 	n.chainManager, err = chains.New(
 		&chains.ManagerConfig{
@@ -1425,7 +1451,7 @@ func (n *Node) initHealthAPI() error {
 		return err
 	}
 
-	n.health, err = health.New(n.Log, healthReg)
+	n.health, err = health.New(n.Log, healthReg, n.reportUnhealthy)
 	if err != nil {
 		return err
 	}